+++ src/main/java/com/finalproject/kdiary/external/client/google/KakaoApiClient.java
@@ -1,14 +1,14 @@
-package com.finalproject.kdiary.external.client.kakao;
+package com.finalproject.kdiary.external.client.google;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
-import com.finalproject.kdiary.external.client.kakao.dto.response.KakaoUserResponse;
+import com.finalproject.kdiary.external.client.google.dto.response.GoogleUserResponse;
 
 @FeignClient(name = "kakaoApiClient", url = "https://kapi.kakao.com")
 public interface KakaoApiClient {
 
     @GetMapping(value = "/v2/user/me")
-    KakaoUserResponse getUserInformation(@RequestHeader(HttpHeaders.AUTHORIZATION) String accessToken);
+    GoogleUserResponse getUserInformation(@RequestHeader(HttpHeaders.AUTHORIZATION) String accessToken);
 }
