+++ src/main/java/com/finalproject/kdiary/external/client/google/KakaoAuthApiClient.java
@@ -1,16 +1,16 @@
-package com.finalproject.kdiary.external.client.kakao;
+package com.finalproject.kdiary.external.client.google;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import com.finalproject.kdiary.external.client.kakao.dto.response.KakaoAccessTokenResponse;
+import com.finalproject.kdiary.external.client.google.dto.response.GoogleAccessTokenResponse;
 
 @FeignClient(name = "kakaoAuthApiClient", url = "https://kauth.kakao.com")
 public interface KakaoAuthApiClient {
 
     @PostMapping(value = "/oauth/token", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-    KakaoAccessTokenResponse getOAuth2AccessToken(
+    GoogleAccessTokenResponse getOAuth2AccessToken(
             @RequestParam("grant_type") String grantType,
             @RequestParam("client_id") String clientId,
             @RequestParam("redirect_uri") String redirectUri,
