+++ src/main/java/com/finalproject/kdiary/external/client/google/dto/response/GoogleAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.finalproject.kdiary.external.client.kakao.dto.response;
+package com.finalproject.kdiary.external.client.google.dto.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -9,12 +9,12 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoAccessTokenResponse {
+public class GoogleAccessTokenResponse {
 
     private String accessToken;
     private String refreshToken;
 
-    public static KakaoAccessTokenResponse of(String accessToken, String refreshToken) {
-        return new KakaoAccessTokenResponse(accessToken, refreshToken);
+    public static GoogleAccessTokenResponse of(String accessToken, String refreshToken) {
+        return new GoogleAccessTokenResponse(accessToken, refreshToken);
     }
 }
