+++ src/main/java/com/finalproject/kdiary/external/client/google/dto/response/GoogleAccount.java
@@ -1,4 +1,4 @@
-package com.finalproject.kdiary.external.client.kakao.dto.response;
+package com.finalproject.kdiary.external.client.google.dto.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -9,8 +9,7 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoUserProfile {
+public class GoogleAccount {
 
-    private String nickname;
-    private String profileImageUrl;
+    private GoogleUserProfile profile;
 }
